backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pypdf
google-genai
passlib[bcrypt]
email-validator
itsdangerous
python-dotenv
pydantic
psycopg2
python-jose[cryptography]
